requirements.txt
django==5.0.4
asgiref==3.8.1
sqlparse==0.5.0
typing-extensions==4.11.0
django_rest_framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
PyJWT==2.8.0
social-auth-app-django==5.4.1
django-templated-mail==1.1.1
social-auth-core==4.5.4
python3-openid==3.2.0
defusedxml==0.8.0
requests==2.31.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
cryptography==42.0.7
cffi==1.16.0
idna==3.7
certifi==2024.2.2
urllib3==2.2.1
charset-normalizer==3.3.2
pycparser==2.22
drf-spectacular==0.27.2
jsonschema==4.22.0
uritemplate==4.1.1
PyYAML==6.0.1
rpds-py==0.18.1
referencing==0.35.1
attrs==23.2.0
jsonschema-specifications==2023.12.1
inflecton==0.5.1
coverage==7.5.1